==> research/01_data_exploration.ipynb <==
DATA_PATH = '../artifacts\data_ingestion\Churn_Modelling.csv'

import pandas as pd 
import numpy as np
pd.options.display.max_rows = 10000

# Read data and print shape
raw_df = pd.read_csv(DATA_PATH)
raw_df.shape

# Display random answer 
# Observations: Multiple answers need to be splitted 
# Reference to the schema needed to understand
raw_df.sample(1).iloc[0]

# Print the general information of the data frame 
raw_df.info()

raw_df.head()

# Get stats for the numerical column
raw_df.describe().round(3)

#### Features:
- Surname: Surname
- CreditScore: Credit score
- Geography: Country (Germany / France / Spain)
- Gender: Gender (Female / Male)
- Age: Age
- Tenure: How many years of customer
- Balance: Balance
- NumOfProducts: No. of Banks product used
- HasCrCard: Credit card status (0 = No, 1 = Yes)
- IsActiveMember: Active membership status (0 = No, 1 = Yes)
- EstimatedSalary: Estimated salary
- Exited: Abandoned or not? (0 = No, 1 = Yes)

-----

----
