==> research/02_data_preprocessing.ipynb <==
DATA_PATH = '../artifacts/data_ingestion/Churn_Modelling.csv'

EXPORT_PATH = '../artifacts/data_preprocessed/1_preprocessed_df.pkl'

# Load packages
import pandas as pd 
import numpy as np
import logging
import pickle
from sklearn.preprocessing import LabelEncoder

-------

# Processing 

## Preprocess Data 

raw_df = pd.read_csv(DATA_PATH)
df = raw_df.copy()

### 1. Counting Null Values

null_values = df.isnull().sum()
null_values

### 2. Counting Duplicates

duplicates = df.duplicated().sum()
duplicates

### 3. Exited Customer Distribution Analysis

values = df['Exited'].value_counts()
values

### 4. Removing Unnecessary Columns

df = df.drop(columns=['RowNumber', 'CustomerId', 'Surname'])
df

### 5. Categorizing Dataset Columns (numerical and nominal types)

numbCol = ['EstimatedSalary', 'Balance', 'CreditScore', 'Age']
nomCol =  ['HasCrCard', 'IsActiveMember', 'Geography', 'Gender', 'NumOfProducts', 'Tenure']

[numbCol,nomCol]

### 6. Outlier Detection Analysis

from scipy import stats

# Define the threshold for outlier detection (the threshold is set to 3 standard deviations from the mean)
threshold = 3

# Create a dictionary to store the count of outliers for each column
outlier_counts = {}

# Iterate over each numerical column
for col in numbCol:

    # Calculate Z-scores for the current column
    z_scores = stats.zscore(df[col])
    
    # Find outliers
    outliers = abs(z_scores) > threshold
    
    # Count the number of outliers for the current column
    outlier_count = outliers.sum()
    
    # Store the count of outliers for the current column in the dictionary
    outlier_counts[col] = outlier_count
    
outlier_counts

##### Insight:
---------------------
We've found outliers in two columns. Before removing them, we need to ensure that it's okay to drop them. 
Because age can be any number, although we set one limit and check for real outliers. This examination is crucial for maintaining data integrity and reliability in subsequent analyses

### 7. Age Data Integrity Check

# Filter the DataFrame for values in the "Age" column greater than 100 and less than 1 
ages_check = len(df[(df['Age'] < 1) | (df['Age'] > 100)])
ages_check

##### Insights
--------------
- We have no real outliers in the Age column


### 8. Credit Score Data Integrity Check

- the credit score typically ranges from 300 to 900

# Filter the DataFrame for values in the "Age" column greater than 900 and less than 300
credit_check = len(df[(df['CreditScore'] < 300) | (df['CreditScore'] > 900)])
credit_check

##### Insight:
------------
- There are no real outliers in the credit score column either

### 9. Converting Estimated Salary and Balance to Integer Data Type

df['EstimatedSalary'] = df['EstimatedSalary'].astype(int)
df['Balance'] = df['Balance'].astype(int)

df['EstimatedSalary'].dtype, df['Balance'].dtype

### 10. Label Encoding for Categorical Data

le = LabelEncoder()
lstforle = ['Geography', 'Gender']
for col in lstforle:
    df[col] = le.fit_transform(df[col])


[df['Gender'].unique(), 
df['Geography'].unique()] 

### 11. Percentage of Churned Customers based on Credit Card Ownership

- check how the column (HasCrCard) relates to the 'Exited' column. This analysis helps us understand how credit card services affect customer retention strategies

credit_churn_percentage = df.groupby('HasCrCard')['Exited'].mean()*100
credit_churn_percentage

##### Insights:
--------------
- credit card services have no impact on customer retention

### 12. Removing Credit Card Ownership Column

df = df.drop('HasCrCard', axis=1)
df

___________

## Visually verify results 

i = df.sample(1).index[0]
print(raw_df['Geography'].iloc[i])
print(df['Geography'].iloc[i])

i = df.sample(1).index[0]
print(raw_df['Gender'].iloc[i])
print(df['Gender'].iloc[i])

# Export Data 

df.to_pickle(EXPORT_PATH)

---

pd.read_pickle(EXPORT_PATH)

df.columns

df.dtypes
