==> research/03_data_validation.ipynb <==
import os

%pwd

# os.chdir("Bank Churn Prediction")
os.chdir("../")

%pwd

DATA_PATH = 'artifacts/data_ingestion/Churn_Modelling.csv'

import pandas as pd

data = pd.read_csv(DATA_PATH)
data.head()

data.info()

data.isnull().sum()

data.shape

data.columns

data.dtypes

#### 4. Update Entity

from dataclasses import dataclass
from pathlib import Path


@dataclass(frozen=True)
class DataValidationConfig:
    root_dir: Path
    unzip_data_dir: Path
    STATUS_FILE: str    
    all_schema: dict
    unzip_dir: Path

#### 5. Update Config manager

import os
os.chdir("../")

from src.BankChurn.constants import *
from src.BankChurn.utils.common import read_yaml, create_directories

class ConfigurationManager:
    def __init__(
        self,
        config_filepath=CONFIG_FILE_PATH,
        params_filepath=PARAMS_FILE_PATH,
        schema_filepath=SCHEMA_FILE_PATH,
    ):
        self.config = None
        self.params = None
        self.schema = None
        self.config = read_yaml(config_filepath)
        self.params = read_yaml(params_filepath)
        self.schema = read_yaml(schema_filepath)
        create_directories([self.config.artifacts_root])


    
    def get_data_validation_config(self) -> DataValidationConfig:
        config = self.config.data_validation
        schema = self.schema.COLUMNS

        create_directories([config.root_dir])

        data_validation_config = DataValidationConfig(
            root_dir=config.root_dir,
            STATUS_FILE=config.STATUS_FILE,
            unzip_data_dir = config.unzip_data_dir,
            all_schema=schema,
            unzip_dir=config.unzip_dir 
            
        )

        return data_validation_config

#### 6. Update Components

import os
import pandas as pd
import pickle
from src.BankChurn import logger
from sklearn.preprocessing import LabelEncoder

class DataValidation:
    def __init__(self, config: DataValidationConfig):
        #self.data_path = data_path
        self.config = config
        logger.info('Initialized DataValidation Class')
        

    def load_data(self):
        logger.info('Loading data from {}'.format(self.config.unzip_data_dir))
        self.df = pd.read_csv(self.config.unzip_data_dir)
        logger.info(f"Loaded data from {self.config.unzip_data_dir}")
        logger.info('Loaded data with shape {}'.format(self.df.shape))
        return self.df

    def preprocess(self):
        """ Preprocess the dataframe """
        self.load_data()

        # Drop unnecessary columns
        columns_to_drop = ['RowNumber', 'CustomerId', 'Surname']
        for col in columns_to_drop:
            if col in self.df.columns:
                self.df = self.df.drop(columns=[col])
                logger.info(f"Dropped column {col}")

        # Convert columns to integer type
        for col in ['EstimatedSalary', 'Balance']:
            if col in self.df.columns:
                self.df[col] = self.df[col].astype(int, errors='ignore')
                logger.info(f"Converted column {col} to integer type")

        # Encode categorical columns using LabelEncoder
        le = LabelEncoder()
        categorical_cols = ['Geography', 'Gender']
        for col in categorical_cols:
            if col in self.df.columns:
                self.df[col] = le.fit_transform(self.df[col])
                logger.info(f"Encoded categorical column {col}")

        # Drop additional column
        if 'HasCrCard' in self.df.columns:
            self.df = self.df.drop('HasCrCard', axis=1)
            logger.info("Dropped column HasCrCard")

        # Export preprocessed data
        validated_data_path = os.path.join('artifacts/data_preprocessed', 'preprocessed_data.pkl')
        os.makedirs(os.path.dirname(validated_data_path), exist_ok=True)
        self.df.to_pickle(validated_data_path , protocol=pickle.HIGHEST_PROTOCOL)
        logger.info(f"Exported preprocessed data to {validated_data_path}")

        return self.df

    def validate_columns(self, preprocessed_data: pd.DataFrame) -> bool:
        try:
            validation_status = True
            all_cols = list(preprocessed_data.columns)
            all_schema = self.config.all_schema.keys()

            for col in all_cols:
                if col not in all_schema:
                    validation_status = False
                    logger.error(f"Column {col} not found in schema")
                else:
                    schema_data_type = self.config.all_schema[col]
                    data_data_type = preprocessed_data[col].dtype

                    if schema_data_type != data_data_type:
                        validation_status = False
                        logger.error(f"Data type mismatch for column {col}: schema type {schema_data_type}, data type {data_data_type}")
            
            with open(self.config.STATUS_FILE, 'w') as f:
                f.write(f"Validation status: {validation_status}")

            return validation_status
        except Exception as e:
            logger.exception(f"Error in validating columns: {e}")
            raise e
        
    
    def validate_data(self):
        preprocessed_data = self.preprocess()
        validation_status = self.validate_columns(preprocessed_data)
        logger.info(f"Validation status: {validation_status}")
        return validation_status

#### 7. Update Pipeline

try:
    config = ConfigurationManager()
    data_validation_config = config.get_data_validation_config()
    data_validation = DataValidation(data_validation_config)
    #preprocessed_data = data_validation.preprocess()
    validation_status = data_validation.validate_data()
    print(f"Validation status: {validation_status}")
except Exception as e:
    logger.error(f"Error occurred: {e}")
    raise e
