==> research/04_data_transformation.ipynb <==
DATA_PATH = '../artifacts/data_preprocessed/1_preprocessed_df.pkl'

# Load packages
import pandas as pd 
import numpy as np
import logging
import pickle
from sklearn.preprocessing import StandardScaler
from imblearn.over_sampling import SMOTE
from sklearn.model_selection import train_test_split

df = pd.read_pickle(DATA_PATH)
df

numbCol = ['EstimatedSalary', 'Balance', 'CreditScore', 'Age']
nomCol =  ['IsActiveMember', 'Geography', 'Gender', 'NumOfProducts', 'Tenure']

### 1. Standardizing Numerical Data with StandardScaler

sc = StandardScaler()
df[numbCol] = sc.fit_transform(df[numbCol])
df

### 2. Splitting Data into Features and Target Variable

x1 = df.drop(columns=['Exited'])
y1 = df['Exited']

x1

### 3. Oversampling Minority Class with SMOTE

y1.value_counts()

over = SMOTE(sampling_strategy=1) # number of samples in the minority class should be equal to the number of samples in the majority class after oversampling

x1_resampled = over.fit_resample(x1, y1)[0].values
y1_resampled = over.fit_resample(x1, y1)[1].values

x1_resampled, y1_resampled

print("Before Over-Sampling \n-----------------------")
print([x1.shape, y1.shape])

print('\n')

print("After Over-Sampling\n-----------------------")
print([x1_resampled.shape, y1_resampled.shape])


x1_resampled

df_resampled = np.concatenate((x1_resampled, y1_resampled.reshape(-1, 1)), axis=1)

df_resampled

### 4. Splitting Oversampled Data for Training and Testing

from sklearn.model_selection import train_test_split

train, test = train_test_split(df_resampled, test_size=0.2, random_state=0)

print(train.shape, test.shape)

----------------------------------------------------

#### Update Entity

from dataclasses import dataclass
from pathlib import Path

@dataclass(frozen=True)
class DataTransformationConfig:
    root_dir: Path
    data_path: Path

import os, sys
%pwd

os.chdir('../')
%pwd

#### Update Config manager

from src.BankChurn.constants import *
from src.BankChurn.utils.common import read_yaml, create_directories

class ConfigurationManager:
    def __init__(
        self,
        config_filepath = CONFIG_FILE_PATH,
        params_filepath = PARAMS_FILE_PATH,
        schema_filepath = SCHEMA_FILE_PATH):

        self.config = read_yaml(config_filepath)
        self.params = read_yaml(params_filepath)
        self.schema = read_yaml(schema_filepath)

        create_directories([self.config.artifacts_root])


    
    def get_data_transformation_config(self) -> DataTransformationConfig:
        config = self.config.data_transformation

        create_directories([config.root_dir])

        data_transformation_config = DataTransformationConfig(
            root_dir=config.root_dir,
            data_path=config.data_path,
        )

        return data_transformation_config



#### Update the components

import os
import pandas as pd
import numpy as np
from src.BankChurn import logger
import pickle
from sklearn.preprocessing import StandardScaler
from imblearn.over_sampling import SMOTE
from sklearn.model_selection import train_test_split

class DataTransformationClass:
    def __init__(self, data_path, config: DataTransformationConfig):
        self.data_path = data_path
        self.config = config
        logger.info('Initialized DataTransformationClass')

    def load_data(self):
        logger.info('Loading data from {}'.format(self.data_path))
        df = pd.read_pickle(self.data_path)
        logger.info('Loaded data with shape {}'.format(df.shape))
        return df

    def preprocess_data(self, df):
        numbCol = ['EstimatedSalary', 'Balance', 'CreditScore', 'Age']
        df[numbCol] = StandardScaler().fit_transform(df[numbCol])
        logger.info('Preprocessed data')
        return df

    def resample_data(self, df):
        x1 = df.copy()
        y1 = df['Exited']
        over = SMOTE(sampling_strategy='minority')
        x1_resampled, y1_resampled = over.fit_resample(x1, y1)
        logger.info('Oversampled data')
        return x1_resampled, y1_resampled

    def split_data(self, df_resampled):
        train, test = train_test_split(df_resampled, test_size=0.2, random_state=0)
        logger.info('Split data into train and test with shapes {} and {}'.format(train.shape, test.shape))
        return train, test

    def save_data(self, train, test):
        os.makedirs('artifacts/data_transformation', exist_ok=True)
        train.to_pickle(os.path.join(self.config.root_dir, "train.pkl"),protocol=pickle.HIGHEST_PROTOCOL)
        test.to_pickle(os.path.join(self.config.root_dir, "test.pkl"),protocol=pickle.HIGHEST_PROTOCOL)
        
    def transform_data(self):
        df = self.load_data()
        preprocessed_data = self.preprocess_data(df)
        resampled_data_x, resampled_data_y = self.resample_data(preprocessed_data)
        train, test = self.split_data(resampled_data_x)
        self.save_data(train, test)

#### Update Pipeline

try:
    DATA_PATH = 'artifacts/data_preprocessed/preprocessed_data.pkl'
    config = ConfigurationManager()
    data_transformation_config = config.get_data_transformation_config()
    data_transformation = DataTransformationClass(data_path=DATA_PATH, config=data_transformation_config)
    data_transformation.transform_data()
except Exception as e:
    raise e

import os
%pwd
os.chdir('../')
%pwd

%pwd
